separa funçoes do main
