+++ funcoes.h
@@ -1,0 +1,11 @@
+// funcoes.h
+#ifndef FUNCOES_H
+#define FUNCOES_H
+
+void testarEstruturasSequenciais();
+void testarEstruturasEncadeadas();
+void testarArvoreBinariaBusca();
+void demonstrarEncapsulamento();
+void demonstracaoIntegrada();
+
+#endif
